Header profile

+++ client/src/App.js
@@ -19,6 +19,7 @@ import {
 } from "./containers/boardContainer";
 import { TrainingContainer } from "./containers/mainContainer";
 import Map from "./pages/Map";
+import MapComponent from "./components/MapComponent";
 import Traininginfo from "./pages/Traininginfo";
 import Mydog from "./pages/Mydog";
 import Mydoginfo from "./pages/Mydoginfo";
@@ -45,8 +46,11 @@ function App() {
     setUserId(res.data);
   };
   const destroySession = async () => {
-    const res = await axios.delete(API_BASE_URL + "/member/logout");
-    setUserId({ isLogin: false });
+    const check = window.confirm("정말로 로그아웃 하시겠습니까?");
+    if (check) {
+      const res = await axios.delete(API_BASE_URL + "/member/logout");
+      setUserId({ isLogin: false });
+    }
   };
   useEffect(() => {
     const checkSession = async () => {
@@ -65,9 +69,12 @@ function App() {
           <Route path="/Login" element={<Login getSession={getSession} />} />
           <Route path="/Register" element={<Register />} />
           <Route path="/MyPage" element={<MyPage />} />
-          <Route path="/quizhome/quiz" element={<Quiz  userId={userId}/>} />
+          <Route path="/quizhome/quiz" element={<Quiz userId={userId} />} />
           <Route path="/quizhome" element={<Quizhome />} />
-          <Route path="/board/page/:pageId" element={<BoardPageContainer userId={userId}/>} />
+          <Route
+            path="/board/page/:pageId"
+            element={<BoardPageContainer userId={userId} />}
+          />
           <Route
             path="/board/:contentId"
             element={<BoardDetailContainer userId={userId} />}
@@ -77,6 +84,7 @@ function App() {
             element={<BoardCreateContainer userId={userId} />}
           />
           <Route path="/Map" element={<Map />} />
+          <Route path="/Map/:query" element={<MapComponent />} />
           <Route path="/training/traininginfo" element={<Traininginfo />} />
           <Route path="/training" element={<TrainingContainer />} />
           <Route path="/mydog" element={<Mydog />} />
